feat: Update get_movie_module.py
Now the function inside the module is more robust.  It searches for specific movie data by using the names of the data such as 'runtime' for movie times. Before data was gathered by indexing specific positions in lists which assumes the structure is the same for all websites. This was seen with marvel's the avengers movie as there is no box office figure on the website. The old function pulled an incorrect data point assuming it was the box office figure.